- Moved snackbar to shell folder - ensured that login/logout now use global snackbar -
