+++ client/src/auth/logout/LogoutSuccess.tsx
@@ -1,5 +1,4 @@
 import React from "react";
-import MuiAlert from "@material-ui/lab/Alert";
 
 export type LogoutSuccessProps = {
   username: string;
@@ -9,9 +8,6 @@ export type LogoutSuccessProps = {
 export function LogoutSuccess(props: LogoutSuccessProps) {
   return (
     <>
-      <MuiAlert elevation={6} variant="filled" severity="success">
-        Du wurdest erfolgreich ausgeloogt.
-      </MuiAlert>
       <p>
         Danke <b>{props.username}</b> fürs Nutzen von Marblecollector!
       </p>
