+++ client/src/parent/ChoreTable.tsx
@@ -23,7 +23,10 @@ import { AddChoreDialog } from "./AddChoreDialog";
 import { useEffect, useState } from "react";
 import { DashboardState } from "../shell/DashboardState";
 import { useDashboardTitle } from "../shell/hooks/DashboardTitleHook";
-import { useInfoNotification, useSuccessNotification } from "../Snackbar";
+import {
+  useInfoNotification,
+  useSuccessNotification,
+} from "../shell/hooks/SnackbarHooks";
 
 const useStyles = makeStyles((theme: Theme) =>
   createStyles({
@@ -47,7 +50,7 @@ export function ChoreTable(): JSX.Element {
 
   const showInfo = useInfoNotification();
   const showSuccess = useSuccessNotification();
-        
+
   const bearerToken = useRecoilValue(AppState.userBearerToken);
   const [snack, setSnackState] = useRecoilState(AppState.snackState);
